+++ common/src/main/java/com/wynntils/screens/wynntilsmenu/WynntilsMenuScreen.java
@@ -9,8 +9,8 @@ import com.wynntils.core.components.Managers;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
 import com.wynntils.features.map.MapFeature;
+import com.wynntils.screens.activities.WynntilsCaveScreen;
 import com.wynntils.screens.base.WynntilsMenuScreenBase;
-import com.wynntils.screens.content.WynntilsCaveScreen;
 import com.wynntils.screens.discoveries.WynntilsDiscoveriesScreen;
 import com.wynntils.screens.guides.WynntilsGuidesListScreen;
 import com.wynntils.screens.lootrun.WynntilsLootrunsScreen;
